+++ crates/ruma-common/src/events/enums.rs
@@ -62,6 +62,15 @@ event_enum! {
         "m.message" => super::message,
         #[cfg(feature = "unstable-msc1767")]
         "m.notice" => super::notice,
+        #[cfg(feature = "unstable-msc3381")]
+        #[ruma_enum(alias = "m.poll.start")]
+        "org.matrix.msc3381.poll.start" => super::poll::start,
+        #[cfg(feature = "unstable-msc3381")]
+        #[ruma_enum(alias = "m.poll.response")]
+        "org.matrix.msc3381.poll.response" => super::poll::response,
+        #[cfg(feature = "unstable-msc3381")]
+        #[ruma_enum(alias = "m.poll.end")]
+        "org.matrix.msc3381.poll.end" => super::poll::end,
         #[cfg(feature = "unstable-msc2677")]
         "m.reaction" => super::reaction,
         "m.room.encrypted" => super::room::encrypted,
@@ -280,6 +289,8 @@ impl AnyMessageLikeEventContent {
             mac::KeyVerificationMacEventContent, ready::KeyVerificationReadyEventContent,
             start::KeyVerificationStartEventContent,
         };
+        #[cfg(feature = "unstable-msc3381")]
+        use super::poll::{end::PollEndEventContent, response::PollResponseEventContent};
 
         match self {
             #[rustfmt::skip]
@@ -325,6 +336,16 @@ impl AnyMessageLikeEventContent {
             Self::Image(ev) => ev.relates_to.clone().map(Into::into),
             #[cfg(feature = "unstable-msc3553")]
             Self::Video(ev) => ev.relates_to.clone().map(Into::into),
+            #[cfg(feature = "unstable-msc3381")]
+            Self::PollResponse(PollResponseEventContent { relates_to, .. })
+            | Self::PollEnd(PollEndEventContent { relates_to, .. }) => {
+                let super::poll::ReferenceRelation { event_id } = relates_to;
+                Some(encrypted::Relation::Reference(encrypted::Reference {
+                    event_id: event_id.clone(),
+                }))
+            }
+            #[cfg(feature = "unstable-msc3381")]
+            Self::PollStart(_) => None,
             Self::CallAnswer(_)
             | Self::CallInvite(_)
             | Self::CallHangup(_)