+++ crates/ruma-common/CHANGELOG.md
@@ -14,6 +14,7 @@ Improvements:
 * All push rules are now considered to not apply to events sent by the user themselves
 * Change `events::relation::BundledAnnotation` to a struct instead of an enum
   * Remove `BundledReaction`
+* Add unstable support for polls (MSC3381)
 
 # 0.9.2